+++ PAF/PAF.psm1
@@ -876,7 +876,7 @@ $oldProtocol = [Net.ServicePointManager]::SecurityProtocol
 $ModuleName = "PAF"
 
 # Get the installed version of the module
-$ModuleVersion = [version]"0.2.4"
+$ModuleVersion = [version]"0.2.5"
 
 # Find the latest version of the module in the PSGallery repository
 $LatestModule = Find-Module -Name $ModuleName -Repository PSGallery