+++ PAF/docs/ReleaseNotes.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 
 - [#9](https://github.com/voytas75/PowershellFramework/issues/9)
+- [#8](https://github.com/voytas75/PowershellFramework/issues/8)
 
 ## [0.2.2] - 2023.08.03