+++ PAF/docs/ReleaseNotes.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 
-- n/a
+- improve performance
 
 ### Changed